pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "lyrabeat"
version = "1.0"
authors = [
  { name="Timo Wendner", email="timo.wendner@gmail.com" },
]
description = "Torch Project to detect Beats from acapella audio files"
readme = "README.md"
requires-python = ">=3.8"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]
dependencies = [
  'PyYAML',
  'tqdm',
  'torch',
  'torchaudio'
]

# [project.urls]
# Homepage = "https://github.com/timowendner/"